refactor(bridge): extract aggregation helpers for issue 386
